Rename picasa_client.rb to google_photo_client.rb
