+++ googlephoto_client.rb
