+++ tokengetter.rb
@@ -1,8 +1,8 @@
 require 'thor'
-require './picasa_client'
+require './googlephoto_client'
 
 class TokenGetter < Thor
-  desc "token", "Get picasa web api token"
+  desc "token", "Get GooglePhoto web api token"
 
   def token
     Pinatra::GooglePhotoClient.new(self)
